Renaming module refactoring for triggered dispatcher refactor
