+++ core/parentchain/block-import-dispatcher/Cargo.toml
@@ -11,7 +11,7 @@ sgx_types = { branch = "master", git = "https://github.com/apache/teaclave-sgx-s
 
 # local dependencies
 itc-parentchain-block-importer = { path = "../block-importer", default-features = false }
-itp-block-import-queue = { path = "../../../core-primitives/block-import-queue", default-features = false }
+itp-import-queue = { path = "../../../core-primitives/import-queue", default-features = false }
 
 # sgx enabled external libraries
 thiserror_sgx = { package = "thiserror", git = "https://github.com/mesalock-linux/thiserror-sgx", tag = "sgx_1.1.3", optional = true }
@@ -32,7 +32,7 @@ default = ["std"]
 std = [
     # local
     "itc-parentchain-block-importer/std",
-    "itp-block-import-queue/std",
+    "itp-import-queue/std",
     # no-std compatible libraries
     "log/std",
     "sp-runtime/std",
@@ -44,7 +44,7 @@ sgx = [
     "sgx_tstd",
     # local
     "itc-parentchain-block-importer/sgx",
-    "itp-block-import-queue/sgx",
+    "itp-import-queue/sgx",
     # sgx enabled external libraries
     "thiserror_sgx",
 ]
