+++ enclave-runtime/src/initialization/global_components.rs
@@ -53,7 +53,7 @@ use itc_tls_websocket_server::{
 	config_provider::FromFileConfigProvider, ws_server::TungsteniteWsServer, ConnectionToken,
 };
 use itp_attestation_handler::IntelAttestationHandler;
-use itp_block_import_queue::ImportQueue;
+use itp_import_queue::ImportQueue;
 use itp_component_container::ComponentContainer;
 use itp_extrinsics_factory::ExtrinsicsFactory;
 use itp_node_api::{
@@ -150,8 +150,14 @@ pub type EnclaveParentchainBlockImporter = ParentchainBlockImporter<
 	EnclaveIndirectCallsExecutor,
 >;
 pub type EnclaveParentchainBlockImportQueue = ImportQueue<SignedParentchainBlock>;
+// TODO: Add this..
+// pub type EnclaveParentchainEventImportQueue = ImportQueue<Vec<Vec<u8>>>;
 pub type EnclaveTriggeredParentchainBlockImportDispatcher =
-	TriggeredDispatcher<EnclaveParentchainBlockImporter, EnclaveParentchainBlockImportQueue>;
+	TriggeredDispatcher<
+		EnclaveParentchainBlockImporter,
+		EnclaveParentchainBlockImportQueue,
+		/* EnclaveParentchainEventImportQueue */
+	>;
 
 pub type EnclaveImmediateParentchainBlockImportDispatcher =
 	ImmediateDispatcher<EnclaveParentchainBlockImporter>;
