+++ sidechain/consensus/common/src/error.rs
@@ -63,7 +63,7 @@ pub enum Error {
 	#[error("Could not import block (number: {0}). A block with this number is already imported (current state block number: {1})")]
 	BlockAlreadyImported(BlockNumber, BlockNumber),
 	#[error("Failed to pop from block import queue: {0}")]
-	FailedToPopBlockImportQueue(#[from] itp_block_import_queue::error::Error),
+	FailedToPopBlockImportQueue(#[from] itp_import_queue::error::Error),
 	#[error("Verification Error: {0}")]
 	VerificationError(its_block_verification::error::Error),
 }
